Move getMatcherListFromString method to utils/alertmanager.ts
