+++ public/app/plugins/panel/alertlist/UnifiedAlertList.tsx
@@ -22,7 +22,7 @@ import { INSTANCES_DISPLAY_LIMIT } from 'app/features/alerting/unified/component
 import { useCombinedRuleNamespaces } from 'app/features/alerting/unified/hooks/useCombinedRuleNamespaces';
 import { useUnifiedAlertingSelector } from 'app/features/alerting/unified/hooks/useUnifiedAlertingSelector';
 import { fetchAllPromAndRulerRulesAction } from 'app/features/alerting/unified/state/actions';
-import { parseMatcher } from 'app/features/alerting/unified/utils/alertmanager';
+import { getMatcherListFromString } from 'app/features/alerting/unified/utils/alertmanager';
 import { Annotation } from 'app/features/alerting/unified/utils/constants';
 import {
   getAllRulesSourceNames,
@@ -33,7 +33,6 @@ import { initialAsyncRequestState } from 'app/features/alerting/unified/utils/re
 import { flattenCombinedRules, getFirstActiveAt } from 'app/features/alerting/unified/utils/rules';
 import { getDashboardSrv } from 'app/features/dashboard/services/DashboardSrv';
 import { DashboardModel } from 'app/features/dashboard/state';
-import { replaceVariables } from 'app/plugins/datasource/prometheus/querybuilder/shared/parsingUtils';
 import { AccessControlAction, useDispatch } from 'app/types';
 import { PromAlertingRuleState } from 'app/types/unified-alerting-dto';
 
@@ -77,22 +76,6 @@ export function UnifiedAlertList(props: PanelProps<UnifiedAlertListOptions>) {
   });
 
   const stateList = useMemo(() => getStateList(props.options.stateFilter), [props.options.stateFilter]);
-
-  function getMatcherListFromString(matcher: string) {
-    //we need to remove curly braces and change the separated-coma format to an array of strings
-    //otherwise the API throws 400
-    if (matcher?.length === 0) {
-      return [];
-    }
-    return matcher
-      ?.replace(/[{}]/g, '')
-      .split(',')
-      .map((value) => value.trim())
-      .map((matcherStr) => {
-        const replacedLabelFilter = replaceVariables(matcherStr);
-        return parseMatcher(replacedLabelFilter);
-      });
-  }
 
   const matcherList = useMemo(
     () => getMatcherListFromString(props.options.alertInstanceLabelFilter),
